move rolling update code to its file
